auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
PyJWT==2.8.0
cryptography==41.0.7